Crate tzdb

tzdb — Time Zone Database

Static time zone information for tz-rs.

This crate provides all time zones found in the Time Zone Database, currently in the version 2022c (released 2022-08-16).

See the documentation for a full list the the contained time zones: https://docs.rs/tzdb/latest/tzdb/time_zone/index.html

Usage examples

// get the system time zone
let time_zone = tzdb::local_tz().unwrap();       // tz::TimeZoneRef<'_>
let current_time = tzdb::now::local().unwrap();  // tz::DateTime

// access by identifier
let time_zone = tzdb::time_zone::europe::KYIV;
let current_time = tzdb::now::in_tz(tzdb::time_zone::europe::KYIV).unwrap();

// access by name
let time_zone = tzdb::tz_by_name("Europe/Berlin").unwrap();
let current_time = tzdb::now::in_named("Europe/Berlin").unwrap();

// names are case insensitive
let time_zone = tzdb::tz_by_name("ArCtIc/LongYeArByEn").unwrap();
let current_time = tzdb::now::in_named("ArCtIc/LongYeArByEn").unwrap();

// provide a default time zone
let current_time = tzdb::now::local_or(tzdb::time_zone::GMT).unwrap();
let current_time = tzdb::now::in_named_or(tzdb::time_zone::GMT, "Some/City").unwrap();

Feature flags

by-name (enabled by default, enabled by local) — enables tz_by_name() to get a time zone at runtime by name

list (enabled by default) — enables [TZ_NAMES] to get a list of all shipped time zones

local (enabled by default) — enables local_tz() to get the system time zone

now (enabled by default) — enables the module [now] to get the current time

binary — make the unparsed, binary tzdata of a time zone available

std (enabled by default) — enable features that need the standard library std

alloc (enabled by default, enabled by std) — enable features that need the standard library alloc

fallback (enabled by default) — compile for unknown target platforms, too

Modules

changelog

Changes between the versions

time_zone

All defined time zones statically accessible

Constants

VERSION

The version of the source Time Zone Database

VERSION_HASH

The SHA512 hash of the source Time Zone Database (using the “Complete Distribution”)

Functions

local_tzlocal

Find the time zone of the current system

tz_by_nameby-name

Find a time zone by name, e.g. "Europe/Berlin" (case-insensitive)